refactor(FolderTable): remove FolderTable and associated features
