+++ packages/admin/cms-admin/src/dam/Table/FolderDataGrid.tsx
@@ -22,11 +22,12 @@ import { DamConfig, DamFilter } from "../DamTable";
 import AddFolder from "../FolderForm/AddFolder";
 import EditFolder from "../FolderForm/EditFolder";
 import { clearDamItemCache } from "../helpers/clearDamItemCache";
+import { isFile } from "../helpers/isFile";
+import { isFolder } from "../helpers/isFolder";
 import DamContextMenu from "./DamContextMenu";
 import { useFileUpload } from "./fileUpload/useFileUpload";
+import { damFolderQuery, damItemsListQuery } from "./FolderDataGrid.gql";
 import * as sc from "./FolderDataGrid.sc";
-import { damFolderQuery, damItemsListQuery } from "./FolderTable.gql";
-import { isFile, isFolder } from "./FolderTableRow";
 import { Footer } from "./footer/Footer";
 import { NameColumn } from "./NameColumn";
 import { TableHead } from "./TableHead";
