--- packages/admin/cms-admin/src/dam/Table/damActions/DamActions.gql.ts
@@ -1,7 +1,0 @@
-import { gql } from "@apollo/client";
-
-export const deleteDamFileMutation = gql`
-    mutation DeleteDamFile($id: ID!) {
-        deleteDamFile(id: $id)
-    }
-`;
