+++ packages/admin/cms-admin/src/dam/helpers/isFile.ts
@@ -1,0 +1,3 @@
+import { GQLDamFileTableFragment, GQLDamFolderTableFragment } from "../../graphql.generated";
+
+export const isFile = (item: GQLDamFileTableFragment | GQLDamFolderTableFragment): item is GQLDamFileTableFragment => item.__typename === "DamFile";
