+++ packages/admin/cms-admin/src/dam/helpers/isFolder.ts
@@ -1,0 +1,4 @@
+import { GQLDamFileTableFragment, GQLDamFolderTableFragment } from "../../graphql.generated";
+
+export const isFolder = (item: GQLDamFileTableFragment | GQLDamFolderTableFragment): item is GQLDamFolderTableFragment =>
+    item.__typename === "DamFolder";
